Add files via upload
This is an explainable AI Parkinson's diagnosis project. We used the 3 Python codes during this project. The "Scoring System" is the code used to generate scores for each audio file. The "Shap Acoustic Feature Explainability" was used to showcase the importance of each acoustic feature in model prediction. Finally, the "MLP + CNN + RNN + MKL" is the code used to analyze files and make a diagnosis.